+++ src/type/schema-types.ts
@@ -1,0 +1,11 @@
+import { ObjectDefinition } from "@meta-system/object-definition";
+
+/**
+ * Duplicated interface from Meta-System
+ */
+export interface SchemaType {
+  name : string;
+  format : ObjectDefinition;
+  dbProtocol : string;
+  identifier : string;
+}